split BCN-stuff into separate file
